Delete README.md
SA / Assignment 5 (20pts)
Attention: This assignment counts double (20pts). It is due on 2017-03-07 8:50am EST
Mancala
• Redo your class diagram with sdmlib.
• Implement all methods to make the game playable from tests.
• Implement at least 6 scenarios as tests and generate corresponding storyboards showing at least
three object diagrams of the object space.
• Create all gui mock-ups for a two player network version of your Mancala game (yes, also the
connect screen)
• Implement a very simple gui (can be text based) approximately matching your mock-ups (can
be simpler). If you use JavaFX or Android for the gui, you can earn up to two PoE.
• Implement the network communication (hint: use json and a simple socket).
• The game needs to be playable via the network.
Submit the git repository link or zipped project with instructions how to run it. This project needs to be
live graded. Your team needs to allocate a 15min slot outside of class with me to present the project to
me to get the grade.